rebpf: use repr(transparent) for the BpfMapDef
Tagging this structure as an explicit wrapper around another structure
allows us to safely consider that a pointer to the BpfMapDef can be
converted to a pointer to the underlying map_def. The default Rust data
layout makes no such promise.

The same result could have been achieved using repr(C), but the
transparent representation conveys the exact semantic of wrapper types,
which repr(C) does not, constraining the compiler further.