rebpf: helpers: un-mut the map reference for bpf_map_lookup_elem
This operation should really be regarded as safe with regards to the
underlying map. Furthermore, the BPF implementation regards this very
argument as a const pointer as well !

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/kernel/bpf/helpers.c?h=v4.19#n37

This should allow us to remove some unsafe blocks from the BPF programs
: the whole point of allowing BPF programs to only have maps as globals
is that accessing those via the intrinsics is supposed to be safe.