rebpf: remove superfluous transmute
Using transmute on bitflags is a bit of an overkill, when one can simply
use the bits() method to access the underlying data. Granted, this
method isn't particularly well documented, but it is listed there :

https://docs.rs/bitflags/1.2.1/bitflags/#methods-1

As a bonus, we don't rely on the hope that the generated bitflags struct
is a transparent wrapper around the u32 bitfield. Of course, it *is*
transparent, but still :-).