maps: rework the traits and macros for better type safety
This patch has actually two components, that are so intertwined that
it wasn't feasible to split them.

The new `map_def` macro is there to completely wrap the map wrapper
definition in a sensible way. To be honest, the reason it got started
was simply that rustfmt systematically removed the empty '<>' in the
macros, causing compilation failures.

The new format allows us to semantically use the generic types as key
and value types, which was very useful for the second component.

The type of the content of a map are supposed to be built into the type
of the map itself. Yet, the previous macros tended towards repeating
some of those types when generating the implementations for the lookup
trait. As I'm working on adding an Update trait with similar signatures,
this could very easily be a source of mistakes. Via the new Map trait,
we can now easily deduce the Key and Value types when we build the map
type, and use them directly in the implementation macros without having
to specify them again :-).