rebpf: use thiserror to implement std::error::Error for our own error type
A proper implementation of std::error::Error is required for our errors
to be handled properly by the various error-handling frameworks out
there that our users might be using. The only one that comes to mind
that is not deprecated is `anyhow`, which basically allows a Rust
application not to care about the error types of its dependencies if it
only wants to bubble them up to the user, but the whole point is to
implement a standard trait for all frameworks to latch onto.

The crate `thiserror` provides some near macros to easily implement the
trait.

This should NOT break the API.