BREAK: remove the `lookup_ref` Lookup method
This API cannot be easily and safely implemented without quite a lot of
work. The core of the problem is that a reference *must* point to a
valid object at all times. Of course, we act under the assumption that a
successful call to the C function will give us valid data, but what
about the data that was there *before* the call? More specifically, what
happens if some of the data has a non-trivial drop implementation?

Either we drop it in place before the call, in which case, if the call
fails, the caller will now have somewhere an object which has already
been dropped, and they won't know it, or we don't, which might produce a
leak of resources in case of non-trivial drop. A third solution would be
to do a copy of the object before the call (see std::pointer::read), and
call drop() on that object if the BPF call succeeds, but in order to do
that, we know need to have the Value type bound to Clone, which would
also be a breaking change.

The easiest solution is thus to simply remove the interface.