BREAK:maps : deport the unsafe operations into layout-specific trait implementations
The idea is that each map knows what its value layout looks like, and
those can vary wildly. It's not reasonable to expect the user of the
crate to know the specifics of the layouts, as this sort of abstraction
seems sort of the idea behind this crate.

Thus, we now have two BPF-specific "pointers" traits, and our libbpf
wrappers use those when appropriate instead of raw pointers or
references. It's up to the trait implementers to ensure that what
they're doing is safe, which allows us to mark the libbpf methods
themselves as safe.

For the CpuArray type, since the Vec type doesn't respect the per-cpu
memory layout, we replace the PerCpuArray Vec<T> value type with our new
PerCpuBuffer<T>. So far, this new type is completely opaque to the
outside, accessors will come in later patches.

For the scalar versions, it turns out the previous implementation was
potentially unsafe and/or leaky if the type T implemented Drop. Using
MaybeUninit allows us to resolve the issue by extracting the actual
objects only if the BPF call succeeded.

Note that we add a dependency to a crate in order to have MaybeUninit
support also in 1.34, which is the version of rustc available in Debian
Buster (the current stable). In 1.36 and later it's only a reexport of
the std type.