BREAK: introduce a marker type to differenciate memory layout
Dependong on the type of BPF map used, there are expectations regarding
the memory layout of its values. This patch introduces a new empty
trait, Layout, as well as two marker implementations, in order to be
able in further patches to enforce via the type system that the proper
memory layouts are used.