map_layout: totally rework the MapLayout trait
The new version gets rid of the LayoutBuffer trait, instead opting to
declare for each buffer "default" buffer types as well as functions to
allocate and convert them. This means that the client code doesn't need
to know about which types are available if they don't want to, they can
simply use the ones that their layout describes.