userspace: change the PerCpu default buffer implementation to a boxed slice
Using a full-blown Vec leaves the user with the possibility to change
the size of the buffer between the allocation and the actual use. This
can lead to panics that would not occur on the site of the bug (when
they resize the buffer) but rather later (when they use the buffer).

A boxed slice simply removes the resizing ability altogether, while
still keeping a lot of the niceties we wanted from Vec.