userspace/maps: derive all the Update and Lookup impl from the MapLayout
By making the Map trait aware of both the layout and the fact that a map
is backed by a file descriptor, we can easily derive default
implementations for the Update and Lookup trait, which in turn allows us
to remove a lot of macro code in favor of plain Rust code. In addition,
the PerCpu variants don't have any specific code anymore, the PerCpu
knowledge being entirely delegated to the MapLayout implementation.